Working oled screen and temperature. Had to use a lighter weight screen library because the pro mini kept running out of memory and crashing

+++ src/scope_temperature.h
@@ -10,12 +10,12 @@ class ScopeTemperature {
 public:
     void begin();
     uint8_t getHumidity();
-    int getTemperature();
-    int getDewpoint();
+    float getTemperature();
+    float getDewpoint();
     bool update();
 
 private:
-    int cToF(int c);
+    float cToF(float c);
 
 };
